64-bits fix

+++ modules/misc/dummy/input.c
@@ -39,7 +39,7 @@
 /*****************************************************************************
  * Access functions.
  *****************************************************************************/
-static ssize_t AccessRead( access_t *p_access, uint8_t *p, int i_size )
+static ssize_t AccessRead( access_t *p_access, uint8_t *p, size_t i_size )
 {
     VLC_UNUSED(p_access);
     memset( p, 0, i_size );
